chat/backend/requirements.txt
fastapi>=0.106.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
PyMuPDF>=1.23.0
Pillow>=10.0.0
pytesseract>=0.3.11
opencv-python-headless>=4.9.0.74
nltk>=3.9.0
scikit-learn>=1.4.0
python-dotenv>=1.0.0
